fix: Relocate SwcToEcuMapping class to SWmapping module (CODING_RULE_STYLE_00008, CODING_RULE_STYLE_00009)
- Move SwcToEcuMapping from SystemTemplate/__init__.py to SWmapping.py
- Remove redundant empty SWmapping/ directory
- Update import in SystemTemplate/__init__.py to include SwcToEcuMapping
- Add ARObject import to SWmapping.py for type annotations

Compliance:
- CODING_RULE_STYLE_00008: Python Package Structure ✓
- CODING_RULE_STYLE_00009: Class Organization per AUTOSAR Mapping ✓

Changes:
- Modified: src/armodel/models/M2/AUTOSARTemplates/SystemTemplate/SWmapping.py
  - Added SwcToEcuMapping class
  - Added ARObject import
- Modified: src/armodel/models/M2/AUTOSARTemplates/SystemTemplate/__init__.py
  - Updated import to include SwcToEcuMapping from SWmapping
  - Removed SwcToEcuMapping class definition
- Deleted: src/armodel/models/M2/AUTOSARTemplates/SystemTemplate/SWmapping/ (empty directory)

Test Results:
- All 2348 tests pass
- SwcToEcuMapping import successful from correct location
- Module location: armodel.models.M2.AUTOSARTemplates.SystemTemplate.SWmapping

Progress:
- Fixed 2 of 54 class location violations (3.7% complete)
- Classes in correct location: 732/1937 (37.8%)
- Remaining violations: 52
