+++ src/armodel/models/M2/AUTOSARTemplates/SystemTemplate/SWmapping.py
@@ -2,9 +2,54 @@
 # It defines mappings between software components and their implementations or partitions
 
 from typing import List
+from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ArObject import ARObject
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.Identifiable import Identifiable
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import RefType
 from armodel.models.M2.AUTOSARTemplates.SystemTemplate.InstanceRefs import ComponentInSystemInstanceRef
+
+
+class SwcToEcuMapping(Identifiable):
+    """
+    Represents the mapping between software components and ECU instances
+    in the system, defining how components are assigned to specific
+    ECUs including hardware element and processing unit references.
+    """
+    def __init__(self, parent: ARObject, short_name: str):
+        super().__init__(parent, short_name)
+
+        self.componentIRefs: List[ComponentInSystemInstanceRef] = []
+        self.controlledHwElementRef: RefType = None
+        self.ecuInstanceRef: RefType = None
+        self.processingUnitRef: RefType = None
+
+    def getComponentIRefs(self):
+        return self.componentIRefs
+
+    def addComponentIRef(self, value):
+        self.componentIRefs.append(value)
+        return self
+
+    def getControlledHwElementRef(self):
+        return self.controlledHwElementRef
+
+    def setControlledHwElementRef(self, value):
+        self.controlledHwElementRef = value
+        return self
+
+    def getEcuInstanceRef(self):
+        return self.ecuInstanceRef
+
+    def setEcuInstanceRef(self, value):
+        self.ecuInstanceRef = value
+        return self
+
+    def getProcessingUnitRef(self):
+        return self.processingUnitRef
+
+    def setProcessingUnitRef(self, value):
+        self.processingUnitRef = value
+        return self
+
 
 
 class SwcToImplMapping(Identifiable):
